requirements.txt
keras==2.2.4
tensorflow-gpu==1.15.0
h5py==2.10.0


easydict==1.6
joblib==0.10.3
numpy==1.12.0
opencv-python==3.2.1
Pillow==4.0.0

Pyyaml==5.3